feat: added error and NaN value handling to useInput
If the parser throws an error useInput will console log the error but
will continue to function. It will just use the last input value for the
time being until a parse succeeds, and then will function as normal.

NaN values are never considered equal. This is usually a good things, as
it keeps mathematical systems from thinking two illegal operations are
equal, `5/0 = Math.asin(1.1)`, but for useInput, I only care if the
higher-level component passed back the parsed value verbatim, not
technicalities of mathematics. So useInput will now treat NaN values as
being equal, and passing back down a NaN value works as intended.