feat: added slider and conventional temp components and fixed useInput
Before useInput would fail to update for anything other than numbers as
`isNaN(notNumber)` is true. There was also an issue with useInput
components not updating if the parsed top-level value was unchanged, as
no state would be set. I now use state for the low value to trigger an
update.