feat: Added standard sliders and updated useInput sliders
I removed the use of `useInput` in the Slider (not plural) because it
actually doesn't need it. Converting an html slider value to an int
is already a bijection, so `useInput` has no real utility there. I also
updated the sliders to use labels "low" and "high" values to be more
intuitive to end-users.