feat: added useLens and updated useInput
`useLens` is about to completely replace `useInput`, but I wanted to
check in my `useInput` changes before doing so. I believe `useLens` to
be a more intuitive api, as it reduces the initial parameters to two,
making it easier to remember. It then returns two functions which mirror
the structure of the parameters and also the `useState` hook. The
parameters to those two functions complete the api. I brought in the npm
package fast-deep-equal as well, since it will allow `useLens` to work
even when the upper-level value is reconstructed for some reason. As
long as the upper-level value is equivilant to the parsed value, the
hook will work. fast-deep-equal is extremely popular, appears to be
well-built, and is tiny, at only ~400 bytes when minified and zipped.